Remove unused parseValueFromString, move stripTrailingSlashes into urls
