+++ bin/generate-cloudfront-config
@@ -5,8 +5,7 @@ const _ = require('lodash');
 const config = require('config');
 
 const routes = require('../controllers/routes');
-const utilities = require('../modules/utilities');
-const urlTools = require('../modules/urls');
+const { generateUrlList,  stripTrailingSlashes } = require('../modules/urls');
 const CF_CONFIGS = require('../config/app/distributions');
 
 // lookup cookies from app config
@@ -56,7 +55,7 @@ const makeBehaviourItem = (origin, path, isPostable, allowQueryStrings, originSe
     // strip trailing slashes
     // fixes /welsh => /welsh/ homepage confusion
     // but doesn't break root/homepage '/' path
-    path = utilities.stripTrailingSlashes(path);
+    path = stripTrailingSlashes(path);
 
     // use all HTTP methods for legacy
     const allowedHttpMethods =
@@ -107,7 +106,7 @@ const makeBehaviourItem = (origin, path, isPostable, allowQueryStrings, originSe
 
 // make a list of every URL we need to serve
 // across all origins
-const urlsToSupport = urlTools.generateUrlList(routes);
+const urlsToSupport = generateUrlList(routes);
 
 // construct array of behaviours from a URL list
 // eg. route them to the relevant origins
