+++ config/cloudfront/live.json
@@ -1402,70 +1402,6 @@
                 "Items": [],
                 "Quantity": 0
             },
-            "QueryString": true
-        },
-        "MaxTTL": 31536000,
-        "PathPattern": "/funding/funding-finder",
-        "SmoothStreaming": false,
-        "DefaultTTL": 86400,
-        "AllowedMethods": {
-            "Items": [
-                "HEAD",
-                "DELETE",
-                "POST",
-                "GET",
-                "OPTIONS",
-                "PUT",
-                "PATCH"
-            ],
-            "CachedMethods": {
-                "Items": [
-                    "HEAD",
-                    "GET"
-                ],
-                "Quantity": 2
-            },
-            "Quantity": 7
-        },
-        "MinTTL": 0,
-        "Compress": false
-    },
-    {
-        "TrustedSigners": {
-            "Enabled": false,
-            "Items": [],
-            "Quantity": 0
-        },
-        "LambdaFunctionAssociations": {
-            "Items": [],
-            "Quantity": 0
-        },
-        "TargetOriginId": "ELB_LIVE",
-        "ViewerProtocolPolicy": "redirect-to-https",
-        "ForwardedValues": {
-            "Headers": {
-                "Items": [
-                    "Accept",
-                    "Host"
-                ],
-                "Quantity": 2
-            },
-            "Cookies": {
-                "Forward": "whitelist",
-                "WhitelistedNames": {
-                    "Items": [
-                        "contrastMode",
-                        "blf-alpha-session",
-                        "_csrf",
-                        "blf-ab-1-50"
-                    ],
-                    "Quantity": 4
-                }
-            },
-            "QueryStringCacheKeys": {
-                "Items": [],
-                "Quantity": 0
-            },
             "QueryString": false
         },
         "MaxTTL": 31536000,
